requirements.txt
fastapi
uvicorn
torch
transformers
pytesseract
Pillow
python-multipart
python-dotenv
pydantic